refactor into a lib
