+++ forc-plugins/forc-doc/tests/fixtures/function_impls/.gitignore
@@ -1,0 +1,2 @@
+out
+target
