+++ forc-plugins/forc-doc/tests/fixtures/function_impls/Forc.toml
@@ -1,0 +1,9 @@
+[project]
+authors = ["Fuel Labs <contact@fuel.sh>"]
+entry = "main.sw"
+license = "Apache-2.0"
+name = "function_impls"
+implicit-std = false
+
+[dependencies]
+std = { path = "../../../../../sway-lib-std" }
