+++ forc-plugins/forc-doc/tests/lib.rs
@@ -1,0 +1,4 @@
+use forc_doc;
+
+#[test]
+fn functions_inside_an_impl() {}
